Created the axe and functionality
Made sure the when the axe is used, it can destroy Trees (using the tag). It also will play an animation of the slashing which I only used a cube since the axe model didn't import correctly. Made it a prefab as well it can be dragged in or even change the art on it!